// src/app/model/student-payload.ts
export class StudentPayload {  
    userName: string | undefined;
    studentId: string | undefined;
    firstName: string | undefined;
    lastName: string | undefined;
    middleName: string | undefined;
    fullName: string | undefined;
    homePhone: string | undefined;
    mobilePhone: string | undefined;
    emailAddress: string | undefined;
    streetAddress1: string | undefined;
    streetAddress2: string | undefined;
    city: string | undefined;
    state: string | undefined;
    country:string | undefined;
    zipCode: string | undefined
    
  }